+++ budgie-app/App.tsx
@@ -1,7 +1,6 @@
 import React from 'react';
 import 'react-native-gesture-handler';
 import { NavigationContainer } from '@react-navigation/native';
-import { createStackNavigator } from '@react-navigation/stack';
 import * as eva from '@eva-design/eva';
 import {
   ApplicationProvider,
@@ -13,44 +12,23 @@ import {
 import { EvaIconsPack } from '@ui-kitten/eva-icons';
 import * as SplashScreen from 'expo-splash-screen';
 import { BottomTabBarProps, createBottomTabNavigator } from '@react-navigation/bottom-tabs';
-import { Expenses } from './src/screens/expenses';
-import { CreateExpense } from './src/screens/expenses/CreateExpense';
 import { Login } from './src/screens/Login';
-import { EditExpense } from './src/screens/expenses/EditExpense';
 import { ToastProvider } from './src/ToastProvider';
 import { Tags } from './src/screens/Tags';
+import { ExpensesStack } from './src/screens/expenses';
 
-const Stack = createStackNavigator();
 const { Navigator, Screen } = createBottomTabNavigator();
 
-// TODO: Split this up
 export type RootStackParamList = {
   'Login': undefined,
   'Expenses': undefined,
-  'CreateExpense': undefined,
-  'EditExpense': { id: string }
+  'Tags': undefined,
 };
 
-const ExpensesStack = () => (
-  <Stack.Navigator
-    headerMode="none"
-  >
-    <Stack.Screen
-      name="Expenses"
-      component={Expenses}
-    />
-    <Stack.Screen
-      name="CreateExpense"
-      component={CreateExpense}
-    />
-    <Stack.Screen
-      name="EditExpense"
-      component={EditExpense}
-    />
-  </Stack.Navigator>
-);
-
-SplashScreen.preventAutoHideAsync().then(() => console.log('prevented')).catch((error) => console.error(error));
+// Show SplashScreen until login state is determined.
+SplashScreen
+  .preventAutoHideAsync()
+  .catch((error) => console.error(error));
 
 const BottomTabBar = ({ navigation, state }: BottomTabBarProps) => (
   <BottomNavigation
